+++ public/data/commentList.json
@@ -1,0 +1,32 @@
+[
+  // {
+  //   "id": 0,
+  //   "nickname": "tiger",
+  //   "description": "부럽다 ㅎㅎ"
+  // },
+  // {
+  //   "id": 1,
+  //   "nickname": "white-bear",
+  //   "description": "다음에 또 놀러와"
+  // },
+  // {
+  //   "id": 2,
+  //   "nickname": "chick",
+  //   "description": "삐약삐약"
+  // },
+  // {
+  //   "id": 3,
+  //   "nickname": "deer",
+  //   "description": "부럽다 ㅎㅎ"
+  // },
+  // {
+  //   "id": 4,
+  //   "nickname": "dolphin",
+  //   "description": "재밌었겠다!"
+  // },
+  // {
+  //   "id": 5,
+  //   "nickname": "chicken",
+  //   "description": "춥지 않아?"
+  // }
+]
