+++ src/pages/jaeminYu/Main/Main.js
@@ -1,51 +1,19 @@
 import Nav from '../../../components/Nav/Nav';
 import React from 'react';
 import '../Main/Main.scss';
-import {
-  FaHeart,
-  FaComment,
-  FaPaperPlane,
-  FaBookmark,
-  FaSmile,
-} from 'react-icons/fa';
-import { useState } from 'react';
-import Comment from './Comment';
+import { useState, useEffect } from 'react';
+import Feed from './Feed';
 
 export const Main = () => {
-  const [userComment, setUserComment] = useState('');
-  const [postBtn, setPostBtn] = useState('postingBtnInActive');
-  const [listElement, setListElement] = useState([]); // Event 3-1 : 댓글 추가 내용을 담기 위한 빈 배열 생성
+  const [feedData, setFeedData] = useState([]);
 
-  // Event 1 : input에 입력 되는 값 저장
-  const CommnetLength = event => {
-    setUserComment(event.target.value);
-  };
-
-  // Event 2 : input 입력 값이 0 을 초과하는 경우 버튼 활성화
-  const CommnetBtnHandler = () => {
-    userComment.length > 0
-      ? setPostBtn('postingBtnActive')
-      : setPostBtn('postingBtnInActive');
-  };
-
-  // Event 3 : 댓글 추가 (게시 글을 클릭하는 경우)
-  function UploadComment() {
-    if (userComment.length > 0) {
-      const newComment = {
-        id: '',
-        nickname: 'raccoons',
-        description: userComment,
-      };
-      // Event 3-2 : 스프레드 문법으로 기존 배열(listElement)을 유지하며 newComment 객체 추가
-      setListElement([...listElement, newComment]);
-      setUserComment('');
-    }
-  }
-
-  // Event 4 : 댓글 추가 (엔터를 누를 경우)
-  const CommentEnter = event => {
-    if (event.key === 'Enter' && userComment.length > 0) UploadComment();
-  };
+  useEffect(() => {
+    fetch('http://localhost:3000/data/feedList.json', { method: 'GET' })
+      .then(res => res.json())
+      .then(data => {
+        setFeedData([...data]);
+      });
+  }, []);
 
   return (
     <div>
@@ -76,102 +44,19 @@ export const Main = () => {
               </li>
             </ul>
           </div>
-          <article id="article">
-            <div className="article-header">
-              <div className="article-thumbnail"></div>
-              <div className="article-user-wrap">
-                <p className="article-user">raccoons</p>
-                <p className="address">북극</p>
-              </div>
-              <i className="fas fa-ellipsis-h"></i>
-            </div>
-            <div className="article-img">
-              <div className="img"></div>
-            </div>
-            <div className="article-icon-wrap">
-              <a href="#" className="icon icon-heart">
-                <FaHeart className="far fa-heart" />
-              </a>
-              <a href="#" className="icon icon-comment">
-                <FaComment className="far fa-comment" />
-              </a>
-              <a href="#" className="icon icon-paper-plane">
-                <FaPaperPlane className="far fa-paper-plane" />
-              </a>
-              <a href="#" className="icon icon-bookmark">
-                <FaBookmark className="far fa-bookmark" />
-              </a>
-            </div>
-            <div className="article-like">
-              <div className="like-thumbnail"></div>
-              <p>
-                <span className="like-user">tiger</span>님{' '}
-                <span className="like-total">
-                  외 <span className="like-counting">29</span>명
-                </span>
-                이 좋아합니다
-              </p>
-            </div>
-            <div className="article-description">
-              <div className="desc-header">
-                <p className="desc-txt">
-                  <a href="#" className="desc-user">
-                    raccoons
-                  </a>
-                  북극에 오로라를 보러 왔다. 하늘이 정말 아름답구나. 다시
-                  태어나면 북극곰으로 태어나고 싶다.
-                </p>
-                <button type="button" className="desc-more-btn">
-                  더 보기
-                </button>
-              </div>
-              <ul className="hashtag">
-                <li>#오로라</li>
-                <li>#하늘</li>
-                <li>#북극</li>
-                <li>#북극곰</li>
-              </ul>
-            </div>
-            <div className="article-comment-record">
-              <p className="comment-btn">
-                댓글&nbsp;<span className="comment-counting">0</span>개 모두
-                보기
-              </p>
-              <ul className="comment-list">
-                {/* Event 3-3 : 업데이트 된 listElement 배열을 map으로 돌린다. (Comment 자식 컴포넌트에 props 전달하여 내용 바꾸기) */}
-                {listElement.map((el, index) => (
-                  <Comment
-                    key={index}
-                    nickname={el.nickname}
-                    description={el.description}
-                  />
-                ))}
-              </ul>
-              <p className="comment_time">1일 전</p>
-            </div>
-            <div className="article-comment-posting">
-              <button type="button" className="smile-button">
-                <FaSmile className="far fa-smile" />
-              </button>
-              <input
-                type="text"
-                placeholder="댓글 달기..."
-                className="posting-area"
-                onChange={CommnetLength}
-                onKeyUp={CommnetBtnHandler}
-                value={userComment}
-                onKeyPress={CommentEnter}
-              />
-              {/* Event 3-4 : input태그에 value 속성값에 userComment를 넣어주어 초기화 값을 적용시킨다. */}
-              <button
-                type="button"
-                className={'posting-button' + ' ' + postBtn}
-                onClick={UploadComment}
-              >
-                게시
-              </button>
-            </div>
-          </article>
+          {feedData.map((el, index) => (
+            <Feed
+              key={index}
+              nickname={el.nickname}
+              location={el.location}
+              miniThumbnail={el.miniThumbnail}
+              img={el.img}
+              description={el.description}
+              freindThumbnail={el.freindThumbnail}
+              freindId={el.freindId}
+              likeCount={el.likeCount}
+            />
+          ))}
         </div>
         <div className="main-right">
           <div className="main-fixed-space">
